Remove some bad appActions and update text
